moves Flask-Login logic to _login module, out of app.py
